Fix NPE when adding illegal child to nameless PdfLayer (MUDetect finding 7).
When a PdfLayer is created without a name, e.g., via PdfLayer.createTitle() or when reading OCProperties, and later a child that has already another parent is added to this layer, the addChild(PdfLayer) method throws a NullPointerException. The is because the name is retrieved and immediately converted to a Unicode string, even if it is null.